terraform: cfg node types and worker node count
This change adds variables to configure the node type of
worker nodes and controller, and to set the number of worker
nodes. The documentation in terraform/ was updated accordingly,
and a template file was added to ease set-up.

Lastly, the change contains a .gitignore that prevents
accidental commit of terraform/variables.auto.tfvars since
that file contains settings for specific clusters, and may
contain secrets.

Signed-off-by: Thilo Fromm <thilo@kinvolk.io>