terraform/cluster.tf: allow to use both SSH public key files and text public keys
Currently, default value for ssh_keys variable is incorrect, as code
underneath expects it to be content of id_rsa.pub file and not the path
to this file.

This commit adds new variable 'ssh_key_files', which expects array of
SSH public key file paths, which is later expanded to the correct
content and merged with old 'ssh_keys' variable.

This gives user flexibility to specify either SSH key or SSH key file or
both if needed.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>