scripts/run_benchmarks.sh: reinstall istio when meshing fails
Turns out re-deploying the application is not sufficient when
istio fails to inject sidecars. Instead, the whole istio installation
is borked; istio needs to be deleted and re-installed in order for
injection to work at all.

This change loops over emojivoto installation / injection for
istio and will delete and re-install istio if emojivoto is not
injected successfully; then delete emojivoto pods after istio
was re-installed to trigger sidecar injection. This has succeeded
in multiple tests to overcome injection failure.

Signed-off-by: Thilo Fromm <thilo@kinvolk.io>